test: move fetch tests to fetch folder
